MODIFY app structure to divide tags app from articles app
