+++ articles/views.py
@@ -1,6 +1,6 @@
 from rest_framework import status, viewsets
 from rest_framework.response import Response
-from rest_framework.exceptions import MethodNotAllowed
+from rest_framework.exceptions import MethodNotAllowed, PermissionDenied
 from rest_framework.permissions import IsAuthenticated
 
 from articles.models import Article, Tag
@@ -50,25 +50,3 @@ class ArticleViewSet(viewsets.ModelViewSet):
 
     def partial_update(self, request, *args, **kwargs):
         raise MethodNotAllowed("PATCH")
-
-
-class TagViewSet(viewsets.ModelViewSet):
-    serializer_class = TagSerializer
-    permission_classes = [IsAuthenticated]
-
-    def get_queryset(self):
-        queryset = Tag.objects.filter(user_id=self.request.user.id)
-        return queryset
-
-    def create(self, request, *args, **kwargs):
-        req_data = {
-            'user': request.user.id,
-            'name': request.data.get('tag_name')
-        }
-
-        tag_serializer = TagSerializer(data=req_data)
-        if tag_serializer.is_valid(raise_exception=True):
-            tag_serializer.save()
-
-        return Response({'msg': 'Tag registered successfully'}, status=status.HTTP_201_CREATED)
-
