+++ doc_saver/settings.py
@@ -46,6 +46,7 @@ INSTALLED_APPS = [
     'drf_spectacular',
     'users',
     'articles',
+    'tags'
 ]
 
 MIDDLEWARE = [
