+++ tags/serializers.py
@@ -1,0 +1,10 @@
+from rest_framework import serializers
+
+from articles.models import Tag
+
+
+class TagSerializer(serializers.ModelSerializer):
+
+    class Meta:
+        model = Tag
+        fields = '__all__'
